requirements-doc.txt
pillow
paver
path.py
cogapp==2.3
sphinx